backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
PyMuPDF==1.24.10
SQLAlchemy==2.0.35
pydantic-settings==2.5.2
httpx==0.27.2
